refactor: split map_loaded test file below 700-line hard cap
map_loaded.rs was 679 lines (over the 500-line soft cap). Split along
the natural seam between fragmentation-level tests and player-creation
/ appearance-ordering tests:

- map_loaded.rs (437 lines): mapLoaded builder, fragmentation, TX window
  fit, packet size, fragment ordering, and cinematic-omission tests
- player_creation.rs (new, 210 lines): build_enter_world_body and
  build_create_player appearance-injection and ordering tests, plus
  the map_loaded first-fragment appearance pin

walk_entity_method_records() moved to tests/mod.rs (pub(super)) since
both sibling test files use it.
